Inactive user notification/requirement.txt
functions-framework==3.*
google-cloud-firestore
firebase_admin
google-cloud-pubsub
google-cloud-secret-manager
openai

